requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
apscheduler==3.10.4

